+++ trunk/lmbench3/src/lat_fifo.c
@@ -15,8 +15,8 @@ char	*id = "$Id$\n";
 #define	F1	"/tmp/lmbench_f1.%d"
 #define	F2	"/tmp/lmbench_f2.%d"
 
-void initialize(void *cookie);
-void cleanup(void *cookie);
+void initialize(iter_t iterations, void *cookie);
+void cleanup(iter_t iterations, void *cookie);
 void doit(iter_t iterations, void *cookie);
 void writer(int wr, int rd);
 
@@ -68,10 +68,12 @@ main(int ac, char **av)
 }
 
 void 
-initialize(void *cookie)
+initialize(iter_t iterations, void *cookie)
 {
 	char	c;
 	state_t * state = (state_t *)cookie;
+
+	if (iterations) return;
 
 	sprintf(state->filename1,F1,getpid());
 	sprintf(state->filename2,F2,getpid());
@@ -110,9 +112,11 @@ initialize(void *cookie)
 }
 
 void 
-cleanup(void * cookie)
+cleanup(iter_t iterations, void * cookie)
 {
 	state_t * state = (state_t *)cookie;
+
+	if (iterations) return;
 
 	signal(SIGCHLD, SIG_IGN);
 	if (state->pid) kill(state->pid, 15);
